# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/search-history/{user_id}": {
      "get": {
        "operationId": "SearchController_GetSearchByUserId",
        "summary": "Get search history by user id",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SearchHistoryResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "search-history"
        ]
      },
      "delete": {
        "operationId": "SearchController_deleteSearchHistory",
        "summary": "Delete search history by user id",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "$ref": "#/components/schemas/DeleteSearchHistoryResponse"
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "search-history"
        ]
      }
    },
    "/search-history": {
      "post": {
        "operationId": "SearchController_createSearch",
        "summary": "Add search history",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSearchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "$ref": "#/components/schemas/SearchHistoryResponse"
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "search-history"
        ]
      }
    },
    "/search-history/tradeable/{user_id}": {
      "get": {
        "operationId": "SearchController_GetTradeableSearchByUserId",
        "summary": "Get tradeable search history by user id",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TradeableSearchHistoryResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "search-history"
        ]
      },
      "delete": {
        "operationId": "SearchController_deleteTradeableSearchHistory",
        "summary": "Delete tradeable search history by user id",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "$ref": "#/components/schemas/DeleteTradeableSearchHistoryResponse"
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "search-history"
        ]
      }
    },
    "/search-history/tradeable": {
      "post": {
        "operationId": "SearchController_addTradeableSearch",
        "summary": "Add tradeable search history",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTradeableSearchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "$ref": "#/components/schemas/TradeableSearchHistoryResponse"
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "search-history"
        ]
      }
    },
...
  },
  "info": {
    "title": "KGI Securities API spec",
    "description": "Internal APIs for KGI",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3057/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "apikey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "scheme": "apikey"
      }
    },
    "schemas": {
...
      "SearchHistoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "user_id": {
            "type": "string",
            "example": "1a2b3c4d5e6f7g8h9i0j"
          },
          "ric": {
            "type": "string",
            "example": "APPL"
          },
          "created_at": {
            "type": "string",
            "example": "2024-07-08T09:56:28.356Z"
          }
        },
        "required": [
          "id",
          "user_id",
          "ric",
          "created_at"
        ]
      },
      "CreateSearchDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "Unique identifier for the user",
            "example": "1a2b3c4d5e6f7g8h9i0j"
          },
          "ric": {
            "type": "string",
            "description": "Search string",
            "example": "APPL"
          }
        },
        "required": [
          "user_id",
          "ric"
        ]
      },
      "DeleteSearchHistoryResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Search history for user [xxx] has been deleted."
          }
        },
        "required": [
          "message"
        ]
      },
      "TradeableSearchHistoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "user_id": {
            "type": "string",
            "example": "1a2b3c4d5e6f7g8h9i0j"
          },
          "product": {
            "type": "string",
            "example": "APPL"
          },
          "created_at": {
            "type": "string",
            "example": "2024-07-08T09:56:28.356Z"
          }
        },
        "required": [
          "id",
          "user_id",
          "product",
          "created_at"
        ]
      },
      "AddTradeableSearchDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "Unique identifier for the user",
            "example": "1a2b3c4d5e6f7g8h9i0j"
          },
          "product": {
            "type": "string",
            "description": "Search string",
            "example": "APPL"
          }
        },
        "required": [
          "user_id",
          "product"
        ]
      },
      "DeleteTradeableSearchHistoryResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Tradeable search history for user [xxx] has been deleted."
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}